nutrisync_adk/requirements.txt
google-genai
google-adk
python-dotenv
supabase
fastapi
uvicorn
pydantic
pytz
httpx
